Revert "merging setup and setupbase"
This reverts commit ee1465ec229701aa948e7a6a61ce73a7661c7df6.
